news_paper/requirements.txt
asgiref==3.7.1
Django==4.2.2
django-filter==23.2
sqlparse==0.4.4
typing_extensions==4.6.1
tzdata==2023.3